Commit temp 2 27.6.2022

+++ src/main/java/com/bookshopweb/dao/OrderItemDAO.java
@@ -57,5 +57,7 @@ public interface OrderItemDAO extends DAO<OrderItem> {
     int countByOrderId(@BindList("ids") List<Long> ids);
 
     @SqlQuery("SELECT name FROM product p JOIN order_item o ON p.id = o.productId WHERE o.orderId = :orderId")
-    List<String> getProductNamesByOrderId(@Bind("orderId") int orderId);
+    List<String> getProductNamesByOrderId(@Bind("orderId") long orderId);
+
+
 }
